Crypto.Noise.Tutorial

Description

Noise is a suite of cryptographic protocols similar in spirit to NaCl's crypto_box, or network solutions like TLS, but simpler, faster, with higher-security elliptic-curve cryptography, and stronger guarantees about deniability and identity hiding.

This protocol has many favorable security and usability properties, including:

Sender forward secrecy

After encryption of a Noise box, only the recipient can decrypt it (the sender cannot).

Deniable

The recipient of a Noise box can authenticate the sender, but cannot produce digitally-signed evidence binding the sender to anything.

Identity hiding

Noise boxes reveal no information about the sender or recipient to a 3rd-party observer.

High speed

Noise usues high-speed curves and ciphers designed by Dan Bernstein.

Padded

Noise ciphertext can be padded to avoid leaking plaintext lengths.

Built on "Encrypt-then-MAC" authenticated encryption

Any tampering with ciphertext will cause the recipient to reject the ciphertext prior to decryption.

Noise pipes are built on Noise boxes and designed for interactive communications, and in addition to the above, Noise pipes offer the following benefits:

Full forward secrecy or full key erasure

Compromise of any long-term private keys never compromises old pipes. But furthermore, compromise of an active endpoint in an ongoinng communication does not compromise prior ciphertexts either - a Noise pipe forgets its chain secrets upon every message.

Resistance to key-compromise impersonation

Even with a compromised private key, the compromised party can still verify other parties' identities in a Noise pipe.

Efficient, encrypted handshakes with short roundtrip

Handshakes allow clients and servers to communicate after only one round trip, offering room for validation checks or certificates.

This package offers:

A high-level box API

Boxes are created using the simple seal and open primitives. Boxes are encrypted, authenticated, and optionally anonymous. Furthermore, boxes are forward secret: only the receiver (identified by the receiving public key) can open them.

A simplistic networking API

Noise pipes can be utilized easily on top of TCP sockets using the connect and serve primitives. This makes it easy to write networking services with transparent encryption support, built on familiar send and recv primitives.

A high-level io-streams API

The networking API internally is built off a high-level API based on io-streams, which makes encrypting data over a pipe as easy as reading/writing to an InputStream or OutputStream. This also integrates with the networking API, making it easy to layer in extra transformations (compression, a high level packet format, etc).

X509 certificate support

This package makes it easy for noise pipes to automatically validate X509 certificates as part of the handshake. Parties can begin exchanging data after only one round trip. In this case, the certificate is offered, which is validated before continuing.

For more information visit https://github.com/trevp/noise/wiki.

Introduction

The noise package defines two sets of APIs: boxes and pipes. Boxes handle standalone messages, and pipes encrypt communication channels.

To begin, a sender and a receiver must create a keypair:

sender(senderPK, senderSK) <- createKeypair receiver(receiverPK, receiverSK) <- createKeypair

Send the public keys around, and keep the private keys safe.

Box API

Boxes are created using seal, and opened using open:

>>> b <- seal (Just sender) receiverPK 32 $ pack "Hello world!"
>>> print $ open receiverSK (Just senderPK) b
Just "Hello world!"

When creating a box, you specify the sending keypair, the receiving public key, the amount of random padding you want (to obscure the plaintext length), and the message. To open it, you specify the secret key of the receiving party, and the public key of the sender.

Attempting to open a box from someone other than the sender will result in failure.

Senders may also be anonymous, where the sender does not specify a long-term key pair:

>>> b <- seal Nothing receiverPK 32 $ pack "Hello world!"
>>> print $ open receiverSK Nothing b
Just "Hello world!"

In the above example, the sender of the box is anonymous without a keypair, and attempting to use a value other than Nothing as the key will error. When the sender is anonymous, they are only identified by a short-term ephemeral key, which is used only once for the corresponding box.

Once you have encrypted a value using seal, it can only be decrypted by the receiving party with the secret key. This property means that boxes are forward secret: once you are done creating them and have 'forgotten' the message, you cannot recover it. Furthermore, boxes are deniable: a recipient of a box can authenticate the sender. But they cannot produce signed evidence binding the sender to anything. Finally, boxes do not produce any evidence of who created them or who the receiver is, and resist tampering with a strong MAC.

network API

io-streams API

Padding

All messages in a noise pipe may optionally be padded to help prevent plaintext length analysis on the encrypted pipe. By default, boxes offer no padding - while pipes offer a small amount of random padding, but only in the initial negotiation phase!

To control the amount of padding set in a Noise box, use the input parameter to seal:

pad <- generateRandomNumberBetween 0 512
b <- seal Nothing receiver pad msg
...

This will encrypt the box and pad it with a random amount of data (between 0 and 512 bytes, assuming a good distribution for the number generator).

To pad the initial boxes that are exchanged in a Pipe handshake, you can set the confPadding parameter of the PipeConf. For example, setting confPadding = 32 will pad the initial handshake message you send with 32 random bytes.

Encrypting an active Noise pipe is a little more work. In essence, you must write a function to transform an OutputStream of this type:

f :: OutputStream PipeMsg -> IO (OutputStream ByteString)

There is an alias for this type - PipeMsgPadder.

The type PipeMsg is a simple message type, pairing a ByteString with an integer - the number of bytes to pad the message containing that buffer. You can think of OutputStream PipeMsg as being equivalent to OutputStream (ByteString, Int)

The simplest transformation is constantPadStream, which pads every message with a constant number of random bytes. This is used internally by default if you don't specify your own transformation:

constantPadStream :: Word32 -> PipeMsgPadder
constantPadStream n inp = makeOutputStream $ \v ->
  case v of
    Nothing -> write Nothing inp
    Just x  -> write (Just (PipeMsg x n)) inp

For example, constantPadStream 0 is a routine which will never pad any outgoing messages.constantPadStream 32 is a routine which will pad every message with 32 random bytes.

By default, a Noise pipe uses constantPadStream 0 internally when you do not specify your own padding routine.

In addition, there is also another routine provided - randomPadStream - which will pad every outgoing message with a randomly selected number of bytes (within the specified range).

Either end of a noise pipe can have their own unique padding routines specified without problem.

To specify the padding routine, simply specify it in the confPadder of your configuration. For example, to pad every message from the client with a random number of bytes in the range of 0 to 32, and pad every message from the server with 64 random bytes, you can write:

let sconf = defaultPipeConfig
              { confPadder = Just (constantPadStream 64)
              }

let cconf = defaultPipeConfig
              { confPadder = Just (randomPadStream 0 32)
              }

Note that padding is, of course, not free: as the bytes are taken from /dev/urandom, in some very unscientific benchmarks, a parameter like (randomPadStream 0 4096) cut the speed of encrypting/decrypting /dev/zero and writing it to dev\null by 50% (200MB/s vs 100MB/s), with the Linux kernel's internal cryptographic routines dominating runtime profiles.

Initial message support

Noise pipes have an advantage over TLS in that only one round trip is needed to start sending data - as part of the handshake, the server and client can exchange data inside boxes.

By default, servers and clients are configured to send 16 bytes of zeros in their boxes, and they both validate the boxes have zeros in the handshake. This can be changed by updating PipeConfig before connecting.

This feature allows you to immediately exchange data (taking full advantage of the early roundtrip for your application), or do more exotic things like certificate validation (see below for X509 support).

Using this support is easy. In the PipeConfig for both parties, you can specify two fields:

 1 - confInitialMsg      :: Maybe (IO ByteString)
 2 - confInitialResponse :: Maybe (ByteString -> IO Bool)

These two fields specify the initial handshake procedure (in the order they occur). By default, the handshake looks like:

Step #1: Client ---- Ephemeral Key ---> Server (BEGIN)

Step 1  ---- Server (Initial server message)

Step 2  ---> Server (Initial client message)

Step #4: Client <---- ciphertext -----> Server (Bidirectional communication)

The initial server message in step #2 is specified with confInitialMsg on the server end, and the client checks if this is valid using confInitialResponse. If it is, the client responds with the result of its own confInitialMsg in step #3, and the server checks this message with its confInitialResponse. If this finishes, then the connection is established in step #4 (see Crypto.Noise.Protocol for details).

Note that even if these parameters are set to Nothing, validation always takes place: if you do not specify any override, or use Nothing explicitly, then the server uses the implicit default: a payload of 16 zero bytes each way, with the validation confirming this.

This means that if you override one of the initial messages, you MUST override the counterpart for the other party. This prevents clients and servers from finishing the handshake if they're incorrectly configured on either end. In other words, this is the default operation if you don't specify your own overrides:

confInitialMsg      = Just $ return (replicate 16 0x0)
confInitialResponse = Just $ x -> return (x == replicate 16 0x0)

X509 support

Differences from the standard

Currently, this package mostly implements the Noise specification faithfully.

The large exception is that we use Curve25519 instead of Curve41417. Curve41417 offers a 200-bit security level, which is really what Noise aims to offer (NaCl crypto_box routines similarly only offer 128-bit security). As a result, keys are only 32 bytes as opposed to 52 bytes.

In the future, this protocol will use Curve41417 by default.

Other notes

Here are some important things to note when using this protocol:

You can't turn off the encryption

There is no way to disable encryption. Furthermore, the primitives should be high-speed enough to not negatively impact large workloads.

No protection against traffic analysis by default

While the initial boxes feature randomized padding by default, the outgoing ciphertext is not padded in any way.. If you want to randomize your traffic a bit, use the randomPadStream transformer in your configuration to automatically pad outgoing traffic. The underlying random number generator is MWC seeded from /dev/urandom, so it should be fast, and pick numbers in the specified range with a good distribution.